PR description:
This PR sets as optional cluster resolution histograms using SIM hits associated to the cluster as mtd truth (instead of MtdSimLayerClusters), in ETL and BTL LocalRecoValidation.
With this update, a reduced time consumption is achieved (tested with TimeMemorySummary on TTbarPU):
- btlLocalRecoValid, from ~1.34s to ~0.05s
- etlLocalRecoValid, from ~18.83s to ~0.05s
Furthermore, ExtraPtEff and ExtraPtEtl2Eff histogram range were updated in MtdTracksHarvester.cc (to fix an issue introduced with PR #45457) and mtdHarvesting_cfg.py was moved to D110 scenario.
Finally, some unused parameters were removed from MtdTracksValidation.cc and vertex time resolution ranges and binning were updated.

PR validation:
The updated validation code has been tested in the CMSSW_14_2_0_pre1 release with the workflows: 
- 29634.0_TTbar_14TeV+2026D110
- 29834.0_TTbar_14TeV+2026D110PU
The new code produces histograms compatible to the current validation.