@cmsbuild 2024-10-03
A new Pull Request was created by  @rdelliga for master.

It involves the following packages:

- Validation/MtdValidation (**dqm**)


@antoniovagnerini, @cmsbuild, @nothingface0, @rvenditti, @syuvivida, @tjavaid can you please review it and eventually sign? Thanks.
@fabiocos, @martinamalberti, @missirol this is something you requested to watch as well.
@antoniovilela, @mandrenguyen, @rappoccio, @sextonkennedy you are the release manager for this.

cms-bot commands are listed <a href="http://cms-sw.github.io/cms-bot-cmssw-cmds.html">here</a>