@antoniovagnerini 2024-10-07
+1
- spurious differences seen from unstable WF `12834.7` (open issue #39803 )